+++ system/database/drivers/sqlite3/sqlite3_driver.php
@@ -75,12 +75,18 @@ class CI_DB_sqlite3_driver extends CI_DB {
 	// --------------------------------------------------------------------
 
 	/**
-	 * Non-persistent database connection
+	 * Create database connection
 	 *
+	 * @param	bool
 	 * @return	object	type SQLite3
 	 */
-	public function db_connect()
+	public function db_connect($persistent = FALSE)
 	{
+		if ($persistent === TRUE)
+		{
+			log_message('debug', "SQLite3 doesn't support persistent connections");
+		}
+
 		try
 		{
 			return ( ! $this->password)
@@ -91,19 +97,6 @@ class CI_DB_sqlite3_driver extends CI_DB {
 		{
 			return FALSE;
 		}
-	}
-
-	// --------------------------------------------------------------------
-
-	/**
-	 * Persistent database connection
-	 *
-	 * @return  object	type SQLite3
-	 */
-	public function db_pconnect()
-	{
-		log_message('debug', 'SQLite3 doesn\'t support persistent connections');
-		return $this->db_connect();
 	}
 
 	// --------------------------------------------------------------------