+++ user_guide_src/source/changelog.rst
@@ -130,6 +130,7 @@ Release Date: Not Released
    -  Added ODBC support for create_database(), drop_database() and drop_table() in :doc:`Database Forge <database/forge>`.
    -  Added PDO support for create_database(), drop_database and drop_table() in :doc:`Database Forge <database/forge>`.
    -  Added unbuffered_row() method for getting a row without prefetching whole result (consume less memory).
+   -  Deprecated the db_pconnect() method - it is now an alias for db_connect(TRUE).
 
 -  Libraries